Move functions now return the state
All move functions in grand now return the simulation state. This is
so that the state does not have to be retrieved again (expensive)
outside grand if it is needed. We no longer want to do dcd reporting
within grand because mdtraj has a bug that means that dcd files cannot
be read by other processes while they are open for writing, meaning we
cannot display the dynamics results while the simulation is still
running. We now do dcd reporting with OpenMM and can make use of the
state as calculated in the grand move function.

(SFT-23453)