Fixed buggy chain assignment algorithm in grand
If you start with a protein with 9 chains, openmm.app.Modeller.add_solvent
adds the solvent as chain '10'. The two-character chain name broke the
code for determining what chain the ghost waters should be in. New code finds
the first unused chain letter and uses that, or chain '_' if no other can be
found.

(SFT-23979)